src/requirements.txt
openai
google-generativeai
meta_ai_api
PyYAML